// src/app/shared/models/cart.interface.ts
export interface IcartItem {
  id: string;
  title: string;
  image: string;
  category: string;
  stock: number;
  size: string;
  color?: string;
  price: number;
  quantity: number;
  totalPrice: number;
}

export interface Icart {
  items: IcartItem[];
  totalPrice: number;
  totalQuantity: number;
}